populate gernere id 6

+++ VideoRental/Models/Movie.cs
@@ -12,7 +12,7 @@ namespace VideoRental.Models
         public DateTime ReleaseDate { get; set; }
         public DateTime AddedDate { get; set; }
         public int NumberInStock { get; set; }
-        public byte GenreId { get; set; }
+        public int GenreId { get; set; }
         /*      
                public Genre Genre { get; set; }*/
     }
